[FUBAR] Adapt messages to use serialize library
Currently, event-store-messaging passes a Hash with underscore cased,
String typed keys to the build class methods of our events/commands. We
have to symbolize these keys in our serializers. We don't have a library
to handle this case currently, so we had to misappropriate the ruby JSON
library inline in the code to handle the special cases.